+++ lib/core/controller.dart
@@ -92,11 +92,11 @@ class CoreController {
   Future<String> setupConfig({
     required SetupParams params,
     required SetupState setupState,
-    VoidCallback? preloadInvoke,
+    FutureOr<void> Function()? preloadInvoke,
   }) async {
-    final res = _interface.setupConfig(params);
+    final res = await _interface.setupConfig(params);
     if (preloadInvoke != null) {
-      preloadInvoke();
+      await Future.sync(preloadInvoke);
     }
     return res;
   }